+++ src/main/kotlin/strongdmm/event/type/service/TriggerInstanceService.kt
@@ -6,12 +6,14 @@ import strongdmm.byond.dmm.TileItem
 import strongdmm.event.Event
 
 abstract class TriggerInstanceService {
-    class GenerateInstancesFromIconStates(body: TileItem, callback: (Unit) -> Unit) : Event<TileItem, Unit>(body, callback)
-    class GenerateInstancesFromDirections(body: TileItem, callback: (Unit) -> Unit) : Event<TileItem, Unit>(body, callback)
+    class GenerateInstancesFromIconStates(body: TileItem) : Event<TileItem, Unit>(body, null)
+    class GenerateInstancesFromDirections(body: TileItem) : Event<TileItem, Unit>(body, null)
 
     class FindInstancePositionsByType(body: Pair<MapArea, String>, callback: (List<Pair<TileItem, MapPos>>) -> Unit) :
         Event<Pair<MapArea, String>, List<Pair<TileItem, MapPos>>>(body, callback)
 
     class FindInstancePositionsById(body: Pair<MapArea, Long>, callback: (List<Pair<TileItem, MapPos>>) -> Unit) :
         Event<Pair<MapArea, Long>, List<Pair<TileItem, MapPos>>>(body, callback)
+
+    class DeleteInstance(body: TileItem) : Event<TileItem, Unit>(body, null)
 }