+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+Features:
+ * Added an option to delete instances in the "Object Panel".
+
 Fixes:
  * Incorrect nudge/dir value in the tile popup after another one was opened.